Refs #21280: Change folder name and delete rtps old folder
