+++ examples/cpp/rtps/CMakeLists.txt
@@ -1,4 +1,4 @@
-# Copyright 2019 Proyectos y Sistemas de Mantenimiento SL (eProsima).
+# Copyright 2024 Proyectos y Sistemas de Mantenimiento SL (eProsima).
 #
 # Licensed under the Apache License, Version 2.0 (the "License");
 # you may not use this file except in compliance with the License.
@@ -12,9 +12,56 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-add_subdirectory(AsSocket)
-add_subdirectory(Registered)
+cmake_minimum_required(VERSION 3.20)
 
-if(SQLITE3_SUPPORT)
-    add_subdirectory(Persistent)
+project(fastdds_rtps_example VERSION 1 LANGUAGES CXX)
+
+# Find requirements
+if(NOT fastcdr_FOUND)
+    find_package(fastcdr 2 REQUIRED)
 endif()
+
+if(NOT foonathan_memory_FOUND)
+    find_package(foonathan_memory REQUIRED)
+endif()
+
+if(NOT fastdds_FOUND)
+    find_package(fastdds 3 REQUIRED)
+endif()
+
+#Check C++11
+include(CheckCXXCompilerFlag)
+if(CMAKE_CXX_COMPILER_ID MATCHES "GNU|Clang")
+    check_cxx_compiler_flag(-std=c++11 SUPPORTS_CXX11)
+    if(NOT SUPPORTS_CXX11)
+        message(FATAL_ERROR "Compiler doesn't support C++11")
+    endif()
+endif()
+
+message(STATUS "Configuring Registered...")
+file(GLOB RTPS_SOURCES_CPP "*.cpp")
+
+add_executable(rtps ${RTPS_SOURCES_CPP})
+target_compile_definitions(rtps PRIVATE
+    $<$<AND:$<NOT:$<BOOL:${WIN32}>>,$<STREQUAL:"${CMAKE_BUILD_TYPE}","Debug">>:__DEBUG>
+    $<$<BOOL:${INTERNAL_DEBUG}>:__INTERNALDEBUG> # Internal debug activated.
+)
+target_link_libraries(rtps fastdds fastcdr foonathan_memory)
+install(TARGETS rtps
+    RUNTIME DESTINATION {DATA_INSTALL_DIR}/fastdds/examples/cpp/rtps/${BIN_INSTALL_DIR}
+    )
+
+# Copy the XML files over to the build directory
+file(GLOB_RECURSE XML_FILES ${CMAKE_CURRENT_SOURCE_DIR}/*.xml)
+# for each xml file detected
+foreach(XML_FILE_COMPLETE_PATH ${XML_FILES})
+    # obtain the file name
+    get_filename_component(XML_FILE ${XML_FILE_COMPLETE_PATH} NAME_WE)
+    # copy the file from src to build folders
+    configure_file(
+        ${XML_FILE_COMPLETE_PATH}                   # from full src path
+        ${CMAKE_CURRENT_BINARY_DIR}/${XML_FILE}.xml # to relative build path
+        COPYONLY)
+    install(FILES ${XML_FILE_COMPLETE_PATH}
+        DESTINATION ${DATA_INSTALL_DIR}/fastdds/examples/cpp/rtps/${BIN_INSTALL_DIR})
+endforeach()
