+++ examples/cpp/rtps/ReaderApp.cpp
@@ -72,10 +72,10 @@ using namespace eprosima::fastdds::rtps;
 namespace eprosima {
 namespace fastdds {
 namespace examples {
-namespace rtps_entities {
+namespace rtps {
 
 ReaderApp::ReaderApp(
-        const CLIParser::rtps_entities_config& config,
+        const CLIParser::rtps_config& config,
         const std::string& topic_name)
     : samples_(config.samples)
     , rtps_participant_(nullptr)
@@ -230,7 +230,7 @@ void ReaderApp::stop()
     terminate_cv_.notify_all();
 }
 
-} // namespace rtps_entities
+} // namespace rtps
 } // namespace examples
 } // namespace fastdds
 } // namespace eprosima
